refactor(_shared/domains/payment): verifyRefundEligibility 추출
`refund_utils.ts` 에서 pure 정책 부분 (`verifyRefundEligibility` + 타입) 을
`_shared/domains/payment/refund_policy.ts` 로 이동. IO + env access 가 필요한
`executeRefund` / `RefundError` 는 `refund_utils.ts` 에 잔존.

- refund_policy.ts — pure 정책 + 6 unit tests
- refund_utils.ts — IO + RefundError 만 보존
- payment-cancel / user-cancel-order — import 경로 분리
- refund_utils_test.ts → refund_policy_test.ts 로 이동 (동일 내용)
- BLUEDOC — refund_policy 등록 (refund_amount.ts 만 예정 잔존)

Constraint: pure / IO 분리 — `_shared/domains/<domain>/` 컨벤션
Rejected: re-export 로 하위 호환 유지 | 잡음 / 정리하면 끝
Confidence: high
Scope-risk: narrow
Directive: refund 정책 분기 추가 시 refund_policy.ts 의 시그니처만 확장. PortOne 호출은 refund_utils.ts 에서만
