+++ supabase/functions/_shared/domains/payment/refund_policy_test.ts
@@ -1,6 +1,8 @@
-// Fix #1235: verifyRefundEligibility unit tests — includes event-started guard
-import { assertEquals } from "@std/assert";
-import { verifyRefundEligibility } from "./refund_utils.ts";
+// _shared/domains/payment/refund_policy_test.ts
+// pure unit tests — mock 0, IO 0
+
+import { assertEquals } from "jsr:@std/assert@1";
+import { verifyRefundEligibility } from "./refund_policy.ts";
 
 const GRACE_HOURS = 2;
 const CUTOFF_DAYS = 7;
@@ -15,8 +17,8 @@ function days(d: number): number {
 
 Deno.test("verifyRefundEligibility - event not started + within grace → eligible", () => {
   const now = new Date();
-  const paidAt = new Date(now.getTime() - ms(1)).toISOString(); // 1시간 전 결제
-  const eventStartTime = new Date(now.getTime() + days(3)).toISOString(); // 3일 후 이벤트
+  const paidAt = new Date(now.getTime() - ms(1)).toISOString();
+  const eventStartTime = new Date(now.getTime() + days(3)).toISOString();
 
   const result = verifyRefundEligibility({
     paidAt,
@@ -31,8 +33,8 @@ Deno.test("verifyRefundEligibility - event not started + within grace → eligib
 
 Deno.test("verifyRefundEligibility - event not started + within cutoff → eligible", () => {
   const now = new Date();
-  const paidAt = new Date(now.getTime() - ms(5)).toISOString(); // 5시간 전 결제 (grace 불합격)
-  const eventStartTime = new Date(now.getTime() + days(10)).toISOString(); // 10일 후 이벤트
+  const paidAt = new Date(now.getTime() - ms(5)).toISOString();
+  const eventStartTime = new Date(now.getTime() + days(10)).toISOString();
 
   const result = verifyRefundEligibility({
     paidAt,
@@ -48,8 +50,8 @@ Deno.test("verifyRefundEligibility - event not started + within cutoff → eligi
 // Fix #1235: 이벤트 시작 후 예매 취소 차단 핵심 케이스
 Deno.test("verifyRefundEligibility - event already started → not eligible", () => {
   const now = new Date();
-  const paidAt = new Date(now.getTime() - ms(1)).toISOString(); // 1시간 전 결제 (grace 합격 조건)
-  const eventStartTime = new Date(now.getTime() - ms(1)).toISOString(); // 1시간 전 이벤트 시작
+  const paidAt = new Date(now.getTime() - ms(1)).toISOString();
+  const eventStartTime = new Date(now.getTime() - ms(1)).toISOString();
 
   const result = verifyRefundEligibility({
     paidAt,
@@ -66,7 +68,7 @@ Deno.test("verifyRefundEligibility - event already started → not eligible", ()
 Deno.test("verifyRefundEligibility - event just started (exactly now) → not eligible", () => {
   const now = new Date();
   const paidAt = new Date(now.getTime() - ms(1)).toISOString();
-  const eventStartTime = now.toISOString(); // 정확히 지금 시작
+  const eventStartTime = now.toISOString();
 
   const result = verifyRefundEligibility({
     paidAt,
@@ -82,8 +84,8 @@ Deno.test("verifyRefundEligibility - event just started (exactly now) → not el
 
 Deno.test("verifyRefundEligibility - event not started + grace expired + cutoff too soon → not eligible", () => {
   const now = new Date();
-  const paidAt = new Date(now.getTime() - ms(5)).toISOString(); // 5시간 전 결제 (grace 불합격)
-  const eventStartTime = new Date(now.getTime() + days(3)).toISOString(); // 3일 후 이벤트 (cutoff 불합격)
+  const paidAt = new Date(now.getTime() - ms(5)).toISOString();
+  const eventStartTime = new Date(now.getTime() + days(3)).toISOString();
 
   const result = verifyRefundEligibility({
     paidAt,
@@ -99,7 +101,7 @@ Deno.test("verifyRefundEligibility - event not started + grace expired + cutoff 
 
 Deno.test("verifyRefundEligibility - paidAt null + event not started + within cutoff → eligible", () => {
   const now = new Date();
-  const eventStartTime = new Date(now.getTime() + days(10)).toISOString(); // 10일 후
+  const eventStartTime = new Date(now.getTime() + days(10)).toISOString();
 
   const result = verifyRefundEligibility({
     paidAt: null,
